[CLEANUP][BREAKING] Remove runtime dep on opcode-compiler
Breaking: rename PartialDefinition to PartialDefinitionImpl so an editor
doesn't pull in the implementation from @glimmer/opcode-compiler when
you wanted the interface in @glimmer/interfaces.

Move template handle/error unwrap utils to @glimmer/util.

Move all debug logging functions to @glimmer/debug.
