+++ packages/@glimmer/debug/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/stack-check';
 export * from './lib/metadata';
 export { opcodeMetadata } from './lib/opcode-metadata';
+export { debug, debugSlice, logOpcode } from './lib/debug';
