+++ packages/@glimmer/integration-tests/lib/components/basic.ts
@@ -10,10 +10,9 @@ import {
   ComponentCapabilities,
 } from '@glimmer/interfaces';
 import { TestComponentDefinitionState } from './test-component';
-import { unreachable, expect } from '@glimmer/util';
+import { unreachable, expect, unwrapTemplate } from '@glimmer/util';
 import { VersionedPathReference, UpdatableRootReference } from '@glimmer/reference';
 import { Tag, CONSTANT_TAG } from '@glimmer/validator';
-import { unwrapTemplate } from '@glimmer/opcode-compiler';
 
 export interface BasicComponentFactory {
   new (): BasicComponent;
