+++ packages/@glimmer/integration-tests/lib/components/emberish-root-view.ts
@@ -4,8 +4,7 @@ import { assertElement, firstElementChild } from '../dom/simple-utils';
 import { UpdatableRootReference } from '@glimmer/reference';
 import { renderJitMain, clientBuilder } from '@glimmer/runtime';
 import { SimpleElement } from '@simple-dom/interface';
-import { assign } from '@glimmer/util';
-import { unwrapTemplate, unwrapHandle } from '@glimmer/opcode-compiler';
+import { assign, unwrapTemplate, unwrapHandle } from '@glimmer/util';
 
 export class EmberishRootView {
   private template: Template;
