+++ packages/@glimmer/util/index.ts
@@ -22,6 +22,7 @@ export { assign, fillNulls, values } from './lib/object-utils';
 export * from './lib/platform-utils';
 export * from './lib/string';
 export * from './lib/immediate';
+export * from './lib/template';
 
 export { default as debugToString } from './lib/debug-to-string';
